Use a more stack efficient version of mapM
The function mapM is notorious for eating up a lot of stack space.
This diff provides a version of mapM for the Maybe monad and the
list traversable which doesn't eat up the stack.

Having a version of mapM which doesn't eat a lot of stack is
particularly helpful when trying to find spaceleaks with the
method described here:
http://neilmitchell.blogspot.com/2015/09/detecting-space-leaks.html